Refactor: Split Blog page into smaller components
Refactor the Blog page into smaller, more manageable components: BlogHeader, BlogList, BlogCard, and BlogDialog. This improves code organization and maintainability.
